requirements.txt
chainlit
langchain
google-generativeai
openai-whisper
sounddevice
scipy
pyttsx3
pandas
playsound==1.2.2
python-dotenv
pydub
ffmpeg-python
gtts
chromadb 
langchain 
openpyxl
pypdf 
langchain_community